tests: removed integ tests and updated unit test for code_correlation
